Query secured

+++ sql/uninstall.php
@@ -29,6 +29,9 @@
  * Maybe the merchant will just try to reset the module
  * but does not want to loose all of the data associated to the module.
  */
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
 $sql = [];
 $sql[] = 'DROP TABLE `' . _DB_PREFIX_ . 'omniversepricing_products`;';
 foreach ($sql as $query) {
